requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path`
#
# TODO: Lock to specific versions

GoogleAppEngineCloudStorageClient
pyasn1
rsa
pusher
authomatic
ansiconv
